use the new package names go 1.16 uses for various ioutil functions

+++ daemon/httpd/handler/file_upload.go
@@ -5,7 +5,6 @@ import (
 	"encoding/hex"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"net/http"
 	"os"
 	"path/filepath"
@@ -74,14 +73,15 @@ func (upload *HandleFileUpload) render(w http.ResponseWriter, r *http.Request, m
 func (upload *HandleFileUpload) periodicallyDeleteExpiredFiles() {
 	for {
 		time.Sleep(FileUploadCleanUpIntervalSec * time.Second)
-		files, err := ioutil.ReadDir(fileUploadStorage)
+		files, err := os.ReadDir(fileUploadStorage)
 		if err != nil {
 			upload.logger.Warning("periodicallyDeleteExpiredFiles", "", err, "failed to read file upload directory")
 			continue
 		}
 		var anyFileExpired bool
-		for _, fileInfo := range files {
-			if fileInfo.ModTime().Before(time.Now().Add(-(FileUploadExpireInSec * time.Second))) {
+		for _, fileEntry := range files {
+			fileInfo, err := fileEntry.Info()
+			if err != nil && fileInfo.ModTime().Before(time.Now().Add(-(FileUploadExpireInSec * time.Second))) {
 				anyFileExpired = true
 				upload.logger.Info("periodicallyDeleteExpiredFiles", "", os.Remove(fileInfo.Name()), "delete expired file")
 			}
